fix(test): remove unnecessary cast in session checkout test
The turn_count field was being assigned with an explicit u32 cast, but the division of two usize values already produces a usize that is implicitly convertible to u32 in this context. Removing the cast makes the code consistent with the surrounding style and avoids a clippy warning about unnecessary type conversion.

Auto-committed-on: macbook

+++ crates/openhuman-core/src/web_chat/session_checkout_tests.rs
@@ -59,7 +59,7 @@ fn write_thread_transcript(workspace_dir: &Path, stem: &str, thread_id: &str, ro
         model: None,
         created: "2026-09-20T15:33:42Z".into(),
         updated: "2026-09-20T15:36:32Z".into(),
-        turn_count: (rows.len() / 2) as u32,
+        turn_count: rows.len() / 2,
         input_tokens: 0,
         output_tokens: 0,
         cached_input_tokens: 0,
